Move integration tests to dedicated directory
